tweak cargo-up
This was getting tags wrong in some cases. For crates in the root of the
repo, the tag should just be the version number and not the package
name.

If I were starting fresh, I'd probably just keep the package name in the
tag, but I use this convention pretty much everywhere at this point, so
just keep it.