system: re-arrange system configuration structure
Basically, for system configuration, we want to check all of that into
source control. Until now, I had just been sporadically maintaining
system configuration on each machine, but this has become unwieldy. It
also guarantees that I will forget to copy something whenever I wipe a
machine.

Given that my Archlinux install on my laptop is FUBAR, I'm trying to
start this process right. The general scheme is:

  $HOME/etc/path/to/config-file-as-directory/{default,krusty,...}

Where `/etc/path/to/config-file-as-directory` is a directory that refers
to a concrete file path in the /etc directory. Inside this directory is
an optional file named `default`, along with other optional files named
by hostname, e.g., `krusty`. If a hostname file is present, then that
file is symlinked to the corresponding config file location. Otherwise
the `default` file is. If no files are present, then no symlinks are
created.

This simple scheme permits per-system configuration without losing track
of anything.