bin: add script for finding abandoned commit message
This somewhat regularly occurs for me when I forget to set the git
author configuration for a repository, which in turn causes gpg to fail
and thus the commit to fail. I used to think the message was gone
forever, but kudos to @henrywallace for discovering that it still exists
in The Greay Beyond!