system: add wifi media script
This adds a simple script to attach and detach from my nas over sshfs.

The configuration isn't quite correct since this shouldn't be present on
systems with a stable network connection and on 'nas' itself. So for
now, we list the wifi systems exhaustively. It would be nicer to set
this script as the default and disable it specifically on 'nas' and
'frink'.

This script used to be installed as part of the 'arch-install' script,
but this is its proper home. Thus, we remove it from the 'arch-install'
script. And in particular, the new version in this commit is a bit more
robust, since it will try to attach to an existing SSH agent. This
permits the 'media' script to run reliably as it uses sshfs in most
environments.