git: set default branch to master
Apparently git is actually changing the default branch to main:
https://lore.kernel.org/git/pull.762.git.1605221038.gitgitgadget@gmail.com/

I see no good reason to complicate my other scripts where 'master'
is hard-coded with conditional logic to determine the default branch
name. At least, not at this time. I've found every argument I've seen
in favor of migrating away from 'master' as a branch name to be rather
unconvincing. Personally, I think it's more likely that we are living
in the midst of a moral panic.