bin: add home automation alerts for high voltage
I just had the Sense installed because we're concerned about high
voltage. Sense doesn't seem to have a good way to monitor this data
unlike wattage, which is quite visible. So we use their undocumented API
to record our own time series and alert us if it gets too high.