bin: quick 'n' dirty script to plot voltage over time
Without a plot of data points, it's really hard to get a sense of what
the voltage has looked like over time.

This is my first time using gnuplot directly. I'm somewhat impressed
with how easy it was to get something decent looking---primarily by
googling my way to a solution---but it is very complex.