bin: more voltage plot fixes
gnuplot doesn't support timezones AFAK, so we have to subtract our
timezone offset manually. Doing this in awk was... weird. I had to use a
special CONVFMT directive in order to convince awk to print the floating
point number in a way that made sense. Without it, its precision was
really really low and printed the result in scientific notation.