home: remove gkrellm
It's a sad day, but it looks like the slow death gkrellm has been dying
has finally completed. It's totally useless in a HiDPI display, and I
can't be bothered to tinker with it. Moreover, some of the plugins I
used seemed to have disappeared with the new. I could salvage them from
another machine, but it just seems like a losing battle.