config/nvim/coc: make rust-analyzer do the work
Amazingly, I've been limping along with a strange chimera setup for Rust
development. I had been using Syntastic with its Cargo plugin to report
on compiler errors in my source code, while using rust-analzyer for
things like goto-definition.

I had been using this for so long that I didn't even realize I was any
more. It's likely that I had configured my setup this way because of
something that wasn't working right with rust-analyzer and/or COC, and
just gave up and went with Syntastic for error reporting. But at some
point recently, my compiler error reporting stopped working. (And prior
to that, I wouldn't get errors in test code, which was annoying but
something I put up with.) After some digging, I (re-)discovered that I
was using this strange chimera and decided to see what happens if I
disabled Syntastic and made rust-analyzer do error reporting. (Hence,
this commit.)

It turns out that it works *amazingly* well. It's quite a bit snappier
than it was before, and it works in test code as well. w00t!