bin/rust: add program to search through SMS backups
It searches through this crazy XML dump (gigabytes big) for text
messages. It doesn't extract every bit of data (particularly media), but
does pull out all of the plain text.

Normally I'd try to just do this sort of thing ad hoc with grep and awk,
but holy hell XML is too annoying to deal with that way. And the data is
just so big that it really helps to have an easy to use presentation
style.