bin: fix terminal-tmux-remote again
I still can't run alacritty in GNOME (wat), and gnome-terminal literally
doesn't start if you give it the '-e' flag. Even though it emits a
warning saying it is deprecated (not removed), and the man page for
gnome-terminal continues to document it without warning.

And apparently alacritty doesn't support the -- syntax.

Sometimes software really sucks.