bin: add script for setting up simple Rust binary
I do this so often that I can't believe I haven't done this yet. I
usually just do `mkcd ~/tmp/play/whatever` and then:

    touch main.rs
    cargo init
    cargo add foo bar quux

But this doesn't account for .gitignore or rustfmt setup. That starts to
add too many manual steps. So let's just put all of this in one place.

And call the script from a shell function so that we can cd ourselves
into the appropriate directory.