nvim: darker color used for Rust doc comments
With the longer lines at Astral, I need to shrink the font size in order
to have side-by-side panes fit on one screen. I can mostly read it
without trouble, but the brighter color of the doc comments did actually
become a little hard to read. So I've darkened the color and made it
bold. It's much better now.