env: add MOZ_USE_XINPUT2=1 to enable scrolling
Why isn't this the default!?!?

I had assumed scrolling wasn't smooth Because Linux or some such. But
then I opened Chrome today for $reasons and realized its scrolling was a
lot better than Firefox's. Cue a quick journey to discover that Firefox
can be nice too, but... you need to ask it to be with an environment
variable. Lol.