bin: add samply scripts
`perf` has been quite finnicky and/or slow lately. So I'm going to
try to switch over to `samply` for most of my profiling needs. The
downside of `samply` is that it works by serving HTTP. So if I'm working
on a remote machine (which I usually do, e.g., my workstation at home
from my laptop in the other room), then I need to do some port
forwarding to make that server accessible to me. This requires agreement
on port numbers and other things. So let's make it simple with some
scripts.