config: tweak alacritty config
The previous theme was I believe carried over from Alacritty's default
in the long ago. Certain colors were very hard to read, such as cyan.
This color isn't used too much, but Cargo's `--help` output recently
changed to use it more. That in turn made it difficult for me to read
and thus provoked me to tweak the theme used by Alacritty.

This is the "github-light" theme from alacritty-theme[1].

[1]: https://github.com/alacritty/alacritty-theme/blob/0764548e7f2d4e0b92ce44b5bc9c92ed71df7f7f/themes/github_light.toml