bin/rust: add new 'freq' program
This program computes frequency distributions of bytes, codepoints,
graphemes or words on arbitrary data.

I've frequently wanted a program like this over the years because it's
useful to know how "common" something is in a corpus. In particular,
since many of the optimizations in `regex` depend on this sort of
thing, it's nice to get more visibility into what real data look like.