nvim: make automatic file reloading more robust
The `autoread` option is laughably inept at this. The docs seem to
suggest it is exactly what I want, but it actually triggers in much more
limited scenarios. Apparently it works much better in gvim.

Anyway, it turns out we can hook into some events and more forcefully
call `checktime`, which will do the reloading logic for us. (For
example, it won't automatically stomp on unsaved changes to the buffer.)

But, it turns out, in order for focus events to get propagated from my
terminal emulator to nvim running in a tmux window, we actually also
need to enable focus events in tmux.

All of this should be the default behavior. It is wild that it isn't.