nvim: switch to my fork of rust-vim
The upstream rust-vim project hasn't had any commits for 2 years. While
I generally try to "live" with how it behaves (and it behaves really
well for the most part), I've grown tired of living with the bug where
`text` code fence blocks completely break syntax highlighting. So I
hacked in a fix for it that I'm satisfied with and have pointing nvim to
it.