bin: mount NAS read-only on media PCs
Murphy is getting older now and learning how to use the keyboards for
the media PCs. This increases the risk that he does some kind of
destructive operation. He could easily delete entire directories of
contents, for example. And I don't keep backups of non-personal media
due to the cost.

So when mounting the NAS on these media PCs, let's do it read-only to
protect the data. I very rarely need to do writable or destructive
operations to the NAS from a media PC. And if that becomes a regular
workflow, I can devise some other means of doing that.