bin: support for power hours
This commit add scripts for facilitating the creation of playlists for
power hours. (A drinking game where you take a single shot of beer
every minute for an hour.)

I had been doing this by hand in the past, but that got annoying. These
two scripts make it more streamlined for the future.

Not that I do this often... Once a year (if that).