bin: make `autobright` a no-op on non-laptops
For some reason, this all of a sudden started letting `batrem`'s
failure cause the entire script to fail, even though it was inside a
shell `if`. What in the world?

Anyway, if we aren't on a laptop, then this script doesn't make sense
anyway. So just bail out in that case.