rust-analyzer: make R-A use one shared target directory
This decouples my R-A sessions with my terminal sessions. And in
particular, removes the constraint that only one can be doing a build at
the same time. This can often become annoying on larger projects where
it's common to save, switch to a terminal and then run tests (or
something else requiring a build). When R-A and my terminal sessions use
the same target directory, it's common for the latter step to be blocked
on acquiring a lock to the target directory because R-A has a lock to it
because of the file save I just did.

This usually didn't matter for smaller projects where R-A was quick
enough to not really notice the delay. But for larger projects (like at
Astral), this matters a lot more.