alias: tweak `ssh_reagent` to use updated SSH agent directory
I haven't been able to find anything about this on the Internet, so I'm
guessing it's a pretty fresh change. It was quite annoying to track this
down. But I guess SSH sockets for agents are no longer stored in `/tmp`
but in `~/.ssh/agent`.

Note that we continue checking the old directory because not all of my
Unix machines are recently updated Archlinux environments.